add group, translation

+++ shopify-embed-app/libs/smsdk/smsdk.api.ts
@@ -56,36 +56,53 @@ export interface SendProps {
 }
 
 const send = async ({ receiver, message, scheduled }: SendProps) => {
-  const response: AxiosResponse<SMSDK.Response> = await axios.post(
-    "https://api.sms.dk/v1/sms/send",
-    {
-      receiver,
-      message,
-      senderName: "BySisters",
-      scheduled: scheduled ? scheduled.toISOString().slice(0, -1) : null,
-    },
-    {
-      headers: {
-        "content-type": "application/json",
-        Authorization: "Bearer 4dcc09f3-68e2-11ed-8524-005056010a37",
+  if (process.env.NODE_ENV === "production") {
+    const response: AxiosResponse<SMSDK.Response> = await axios.post(
+      "https://api.sms.dk/v1/sms/send",
+      {
+        receiver,
+        message,
+        senderName: "BySisters",
+        scheduled: scheduled ? scheduled.toISOString().slice(0, -1) : null,
       },
-    }
-  );
+      {
+        headers: {
+          "content-type": "application/json",
+          Authorization: "Bearer 4dcc09f3-68e2-11ed-8524-005056010a37",
+        },
+      }
+    );
+    return response.data;
+  }
 
-  return response.data;
+  return {
+    status: "success",
+    result: {
+      batchId: "bb3ddc53-8969-6d90-de09-134d570c28c1",
+    },
+  };
 };
 
 const cancel = async (batchId: string) => {
-  const response = await axios.delete(
-    `https://api.sms.dk/v1/sms/delete?batchId=${batchId}`,
-    {
-      headers: {
-        "content-type": "application/json",
-        Authorization: "Bearer 4dcc09f3-68e2-11ed-8524-005056010a37",
-      },
-    }
-  );
-  return response;
+  if (process.env.NODE_ENV === "production") {
+    const response = await axios.delete(
+      `https://api.sms.dk/v1/sms/delete?batchId=${batchId}`,
+      {
+        headers: {
+          "content-type": "application/json",
+          Authorization: "Bearer 4dcc09f3-68e2-11ed-8524-005056010a37",
+        },
+      }
+    );
+    return response;
+  }
+
+  return {
+    status: "success",
+    result: {
+      batchId: "bb3ddc53-8969-6d90-de09-134d570c28c1",
+    },
+  };
 };
 
 export default { send, cancel };
